Add sdk Dockerfile templates
