+++ eng/dockerfile-templates/sdk/Dockerfile.ltsc2016
@@ -1,10 +1,28 @@
 # escape=`
 
 ARG REPO=mcr.microsoft.com/dotnet/framework/runtime
-FROM $REPO:4.8-windowsservercore-ltsc2016
+FROM $REPO:{{PRODUCT_VERSION}}-windowsservercore-ltsc2016
+{{if PRODUCT_VERSION = "3.5"
+:
+# Install .NET 4.8 Fx
+RUN powershell -Command `
+        $ProgressPreference = 'SilentlyContinue'; `
+        [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12; `
+        Invoke-WebRequest `
+            -UseBasicParsing `
+            -Uri https://download.visualstudio.microsoft.com/download/pr/7afca223-55d2-470a-8edc-6a1739ae3252/abd170b4b0ec15ad0222a809b761a036/ndp48-x86-x64-allos-enu.exe `
+            -OutFile dotnet-framework-installer.exe `
+    && .\dotnet-framework-installer.exe /q `
+    && del .\dotnet-framework-installer.exe `
+    && powershell Remove-Item -Force -Recurse ${Env:TEMP}\*
 
+# ngen .NET Fx
+RUN \Windows\Microsoft.NET\Framework64\v2.0.50727\ngen uninstall "Microsoft.Tpm.Commands, Version=10.0.0.0, Culture=Neutral, PublicKeyToken=31bf3856ad364e35, processorArchitecture=amd64" `
+    && \Windows\Microsoft.NET\Framework64\v4.0.30319\ngen update `
+    && \Windows\Microsoft.NET\Framework\v4.0.30319\ngen update
+}}
 # Install NuGet CLI
-ENV NUGET_VERSION=4.9.4
+ENV NUGET_VERSION={{VARIABLES[cat("nuget|", OS_VERSION_NUMBER, "|version")]}}
 RUN mkdir "%ProgramFiles%\NuGet\latest" `
     && powershell -Command `
         $ProgressPreference = 'SilentlyContinue'; `
@@ -15,7 +33,7 @@ RUN mkdir "%ProgramFiles%\NuGet\latest" `
             -OutFile $Env:ProgramFiles\NuGet\nuget.exe; `
         Invoke-WebRequest `
             -UseBasicParsing `
-            -Uri https://dist.nuget.org/win-x86-commandline/v5.8.0/nuget.exe `
+            -Uri https://dist.nuget.org/win-x86-commandline/v{{VARIABLES["nuget|version"]}}/nuget.exe `
             -OutFile $Env:ProgramFiles\NuGet\latest\nuget.exe
 
 # Install VS components
@@ -26,7 +44,7 @@ RUN `
         [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12; `
         Invoke-WebRequest `
             -UseBasicParsing `
-            -Uri https://download.visualstudio.microsoft.com/download/pr/5f914955-f6c7-4add-8e47-2e090bdc02fa/8fd51571a9aff13c4381037ccb7559a3bb3e6527cbacf4816068e711dfe824f4/vs_TestAgent.exe `
+            -Uri {{VARIABLES["vs|testAgentUrl"]}} `
             -OutFile vs_TestAgent.exe `
     && start /w vs_TestAgent.exe --quiet --norestart --nocache --wait `
     && powershell -Command "if ($err = dir $Env:TEMP -Filter dd_setup_*_errors.log | where Length -gt 0 | Get-Content) { throw $err }" `
@@ -37,14 +55,14 @@ RUN `
         $ProgressPreference = 'SilentlyContinue'; `
         Invoke-WebRequest `
             -UseBasicParsing `
-            -Uri https://download.visualstudio.microsoft.com/download/pr/5f914955-f6c7-4add-8e47-2e090bdc02fa/bc60b587587d10eacc50fe99a6f3d57a81749cb55c904225892d1d165f01a913/vs_BuildTools.exe `
+            -Uri {{VARIABLES["vs|buildToolsUrl"]}} `
             -OutFile vs_BuildTools.exe `
-    && start /w vs_BuildTools.exe ^ `
-        --add Microsoft.VisualStudio.Workload.MSBuildTools ^ `
-        --add Microsoft.VisualStudio.Workload.NetCoreBuildTools ^ `
-        --add Microsoft.Net.Component.4.8.SDK ^ `
-        --add Microsoft.Component.ClickOnce.MSBuild ^ `
-        --add Microsoft.VisualStudio.Component.WebDeploy ^ `
+    && start /w vs_BuildTools.exe @^ `
+        --add Microsoft.VisualStudio.Workload.MSBuildTools @^ `
+        --add Microsoft.VisualStudio.Workload.NetCoreBuildTools @^ `
+        --add Microsoft.Net.Component.4.8.SDK @^ `
+        --add Microsoft.Component.ClickOnce.MSBuild @^ `
+        --add Microsoft.VisualStudio.Component.WebDeploy @^ `
         --quiet --norestart --nocache --wait `
     && powershell -Command "if ($err = dir $Env:TEMP -Filter dd_setup_*_errors.log | where Length -gt 0 | Get-Content) { throw $err }" `
     && del vs_BuildTools.exe `
@@ -73,7 +91,7 @@ RUN powershell -Command `
         [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12; `
         Invoke-WebRequest `
             -UseBasicParsing `
-            -Uri https://dotnetbinaries.blob.core.windows.net/dockerassets/MSBuild.Microsoft.VisualStudio.Web.targets.2020.11.zip `
+            -Uri {{VARIABLES["vs|webTargetsUrl"]}} `
             -OutFile MSBuild.Microsoft.VisualStudio.Web.targets.zip; `
         Expand-Archive MSBuild.Microsoft.VisualStudio.Web.targets.zip -DestinationPath \"${Env:ProgramFiles(x86)}\Microsoft Visual Studio\2019\BuildTools\MSBuild\Microsoft\VisualStudio\v16.0\" `
     && del MSBuild.Microsoft.VisualStudio.Web.targets.zip
@@ -93,7 +111,7 @@ RUN powershell " `
     $ErrorActionPreference = 'Stop'; `
     $ProgressPreference = 'SilentlyContinue'; `
     [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12; `
-    @('4.0', '4.5.2', '4.6.2', '4.7.2', '4.8') `
+    @@('4.0', '4.5.2', '4.6.2', '4.7.2', '4.8') `
     | %{ `
         Invoke-WebRequest `
             -UseBasicParsing `
