--- eng/nuget-info.json
@@ -1,31 +1,0 @@
-[
-  {
-    "nugetClientVersion": "4.9.4",
-    "nugetClientLatestVersion": "5.8.0",
-    "osVersions": [
-      "windowsservercore-ltsc2016",
-      "windowsservercore-ltsc2019"
-    ]
-  },
-  {
-    "nugetClientVersion": "5.3.1",
-    "nugetClientLatestVersion": "5.8.0",
-    "osVersions": [
-      "windowsservercore-1909"
-    ]
-  },
-  {
-    "nugetClientVersion": "5.5.1",
-    "nugetClientLatestVersion": "5.8.0",
-    "osVersions": [
-      "windowsservercore-2004"
-    ]
-  },
-  {
-    "nugetClientVersion": "5.7.0",
-    "nugetClientLatestVersion": "5.8.0",
-    "osVersions": [
-      "windowsservercore-20H2"
-    ]
-  }
-]
