+++ manifest.versions.json
@@ -1,0 +1,15 @@
+{
+    "variables": {
+      "nuget|ltsc2016|version": "4.9.4",
+      "nuget|ltsc2019|version": "4.9.4",
+      "nuget|1909|version": "5.3.1",
+      "nuget|2004|version": "5.5.1",
+      "nuget|20H2|version": "5.7.0",
+      "nuget|version": "5.8.0",
+  
+      "vs|version": "16.8",
+      "vs|testAgentUrl": "https://download.visualstudio.microsoft.com/download/pr/5f914955-f6c7-4add-8e47-2e090bdc02fa/8fd51571a9aff13c4381037ccb7559a3bb3e6527cbacf4816068e711dfe824f4/vs_TestAgent.exe",
+      "vs|buildToolsUrl": "https://download.visualstudio.microsoft.com/download/pr/5f914955-f6c7-4add-8e47-2e090bdc02fa/bc60b587587d10eacc50fe99a6f3d57a81749cb55c904225892d1d165f01a913/vs_BuildTools.exe",
+      "vs|webTargetsUrl": "https://dotnetbinaries.blob.core.windows.net/dockerassets/MSBuild.Microsoft.VisualStudio.Web.targets.2020.11.zip"
+    }
+  }
